- Move the Business Logic into custom hook
